RAG_APP1/requirements.txt
boto3
awscli
langchain-community
pypdf
streamlit
faiss-cpu
chardet
#pip install -qU langchain_community pypdf
#what is -qU ?